+++ db/migrate/20260101000017_add_reflections_seen_at_to_bible270_readers.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddReflectionsSeenAtToBible270Readers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :bible270_readers, :reflections_seen_at, :datetime
+  end
+end